Upgrade JS SDK and Webpack to latest versions
- Updated JS SDK to latest NPM v3.
- Updated Webpack to latest NPM v5.
- Tested the demo working through all the customer and agent steps
following the README. Verified video works and meeting is connected in both
the agent and customer application as intended.
- Video and audio stops when leaving/ending video meeting through either side.
- Updated README to clarify and callout to use Chrome browser for agent side,
due to certificate issues. For now, keep using Chrome and fix the
firefox separately.
- Remove `babel` and update all the dev and actual dependencies to latest
to fix all the security vulnerabilities.
- Add simple mute/unmute functionality to avoid audio echoing. Currently, there was
no way to stop echoing. Had to use simple buttons in customer side as there
is issue with SVG bundling in customer application even if same setup is followed
like we have in the agent application.
- Add `package-lock.json` to GitHub which not committed.
- Remove unused icons such as content sharing, speaker which were mostly for
showcase and were not handling any functionality.